achievements

+++ src/routes/caught.ts
@@ -1,5 +1,6 @@
 import express, { Request, Response } from 'express';
 import { Caught } from '@/models/caught';
+import { Achievement } from '@/models/achievement';
 const router = express.Router();
 
 router.get('/api/caught', async (req: Request, res: Response) => {
@@ -9,22 +10,43 @@ router.get('/api/caught', async (req: Request, res: Response) => {
 });
 
 router.post('/api/caught', async (req: Request, res: Response) => {
-	const { authId, ueid, critterType } = req.body;
-
-	const createdCaught = await Caught.create({
+	const { authId, ueid, critterType, value } = req.body;
+	const achievement = await Achievement.findOne({ ueid });
+	const newCaught = await Caught.create({
 		authId,
 		ueid,
 		critterType,
+		value,
 		active: true,
 	});
+	const createdCaught = [newCaught];
+	if (value && achievement?.sequential) {
+		for (let i = 0; i < value; i++) {
+			const sequentialCaught = await Caught.create({
+				authId,
+				ueid,
+				critterType,
+				value: i + 1,
+				active: true,
+			});
+			createdCaught.push(sequentialCaught);
+		}
+	}
 	return res.status(201).send(createdCaught);
 });
 
 router.delete('/api/caught', async (req: Request, res: Response) => {
-	const { authId, ueid } = req.body;
-
-	await Caught.deleteMany({ authId, ueid });
-	return res.status(200).send('Deleted successfully');
+	const { authId, ueid, value } = req.body;
+	const achievement = await Achievement.findOne({ ueid });
+	await Caught.deleteMany({ authId, ueid, value });
+	if (achievement?.sequential) {
+		await Caught.deleteMany({ authId, ueid });
+	}
+	return res.status(200).send({
+		ueid,
+		sequential: achievement?.sequential,
+		value,
+	});
 });
 
 export { router as caughtRouter };
